Adding middleware to routes & create comment req routes

+++ routes/web.php
@@ -7,39 +7,49 @@ use Illuminate\Support\Facades\Route;
 //     return view('welcome');
 // });
 
-// Show admin login form
-Route::get('/admin/login', [AdminController::class, 'login']);
 
 //  Show admin dashbord page
-Route::get('/admin', [AdminController::class, 'index']);
+Route::get('/admin', [AdminController::class, 'index'])->middleware('adminAuth');
+
+// Show admin login form
+Route::get('/admin/login', [AdminController::class, 'login'])->middleware('guest');
 
 // authtenticate admin
-Route::post('/admin/auth', [AdminController::class, 'auth']);
+Route::post('/admin/auth', [AdminController::class, 'auth'])->middleware('guest');
 
 // Log out admin
-Route::post('/admin/logout', [AdminController::class, 'logout']);
+Route::post('/admin/logout', [AdminController::class, 'logout'])->middleware('adminAuth');
 
 // Show admin edit username & password form
-Route::get('/admin/change_password', [AdminController::class, 'edit']);
+Route::get('/admin/change_password', [AdminController::class, 'edit'])->middleware('adminAuth');
 
 // Update admin username & password
-Route::put('/admin/update', [AdminController::class, 'update']);
+Route::put('/admin/update', [AdminController::class, 'update'])->middleware('adminAuth');
 
 // Show admin resturents tags page
-Route::get('/admin/resturents_tags', [AdminController::class, 'resturentsTagsIndex']);
+Route::get('/admin/resturents_tags', [AdminController::class, 'resturentsTagsIndex'])->middleware('adminAuth');
 
 // Store admin resturents tag in database
-Route::post('/admin/resturentTagStore', [AdminController::class, 'resturentsTagsStore']);
+Route::post('/admin/resturentTagStore', [AdminController::class, 'resturentsTagsStore'])->middleware('adminAuth');
 
 // Delete admin resturents tag
-Route::delete('/admin/resturents_tags/{resturentTag}', [AdminController::class, 'deleteResTag']);
+Route::delete('/admin/resturents_tags/{resturentTag}', [AdminController::class, 'deleteResTag'])->middleware('adminAuth');
 
 // Show admin food tags page
-Route::get('/admin/food_tags', [AdminController::class, 'foodTagsIndex']);
+Route::get('/admin/food_tags', [AdminController::class, 'foodTagsIndex'])->middleware('adminAuth');
 
 // Store admin food tag in database
-Route::post('/admin/foodTagStore', [AdminController::class, 'foodTagsStore']);
+Route::post('/admin/foodTagStore', [AdminController::class, 'foodTagsStore'])->middleware('adminAuth');
 
 // Delete admin food tag
-Route::delete('/admin/food_tags/{foodTag}', [AdminController::class, 'deleteFoodTag']);
+Route::delete('/admin/food_tags/{foodTag}', [AdminController::class, 'deleteFoodTag'])->middleware('adminAuth');
+
+// Show admin delete comments page
+Route::get('/admin/comments_delete', [AdminController::class, 'commentsDelIndex'])->middleware('adminAuth');
+
+// Keep comment 
+Route::post('/admin/comment_keep/{comment}', [AdminController::class, 'keepComment'])->middleware('adminAuth');
+
+// Delete comment as seller requested
+Route::delete('/admin/comment_deleteing/{comment}', [AdminController::class, 'deleteComment'])->middleware('adminAuth');
